backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-multipart==0.0.9
pdfplumber==0.11.0
pypdfium2==4.30.0
openpyxl==3.1.5
anthropic==0.40.0
Pillow==10.4.0
python-dotenv==1.0.1
stripe==10.0.0
httpx==0.27.0
